* merge upload-log and create-thumbnail to save operator

* tag v0.2.4

+++ chunkflow/operators/block_inference/frameworks/identity_patch_inference_engine.py
@@ -22,7 +22,7 @@ class IdentityPatchInferenceEngine(PatchInferenceEngine):
         """
         patch = self._reshape_patch(patch)
 
-        output = np.copy(patch)
+        output = np.copy(patch).astype(np.float32)
         if self.num_output_channels > 1:
             output = np.repeat(output, self.num_output_channels, axis=1)
